requirements.txt
streamlit>=1.28.0
google-genai>=1.61.0
sentence-transformers>=2.2.0
rank-bm25>=0.2.2
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
